+++ ProyectoJavaCERO/CERO/Persistencia/src/main/java/Entidades/Alumno.java
@@ -1,7 +1,4 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
+
 package Entidades;
 
 import java.time.LocalDate;
@@ -11,7 +8,7 @@ import java.time.LocalDate;
  * @author Usuario
  */
 public class Alumno {
-    private Long codigo;
+    private Integer codigo;
     private String apellidoPaterno;
     private String apellidoMaterno;
     private String nombre;
@@ -19,7 +16,7 @@ public class Alumno {
     private LocalDate fechaNacimiento;
     private String correoElectronico;
 
-    public Alumno(Long codigo, String apellidoPaterno, String apellidoMaterno, String nombre, String telefono, LocalDate fechaNacimiento, String correoElectronico) {
+    public Alumno(Integer codigo, String apellidoPaterno, String apellidoMaterno, String nombre, String telefono, LocalDate fechaNacimiento, String correoElectronico) {
         this.codigo = codigo;
         this.apellidoPaterno = apellidoPaterno;
         this.apellidoMaterno = apellidoMaterno;
